vast-backup-20251015-115024/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
addict==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
bash_kernel==0.10.0
basicsr==1.4.2
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.4
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
conda-pack==0.8.1
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
Cython==3.1.4
dbus-python==1.2.18
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.35.1
distro==1.7.0
easydict==1.13
exceptiongroup==1.3.0
executing==2.2.1
facexlib==0.3.0
fastapi==0.119.0
fastjsonschema==2.21.2
filelock==3.19.1
filetype==1.2.0
filterpy==1.4.5
firebase_admin==7.1.0
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
future==1.0.0
gdown==5.2.0
gfpgan==1.3.8
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.35.3
humanize==4.13.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
importlib-metadata==4.6.4
incremental==24.7.2
iniconfig==2.1.0
insightface==0.7.3
iotop==0.6
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
iterable-io==1.0.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-archive==3.4.0
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==23.5.0
kiwisolver==1.4.9
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
llvmlite==0.45.1
lmdb==1.7.4
lpips==0.1.4
magic-wormhole==0.20.0
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
mistune==3.1.4
ml_dtypes==0.5.3
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbzip==0.1.0
nest-asyncio==1.6.0
networkx==3.3
notebook==7.4.5
notebook_shim==0.2.4
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
onnx==1.19.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
pip==25.2
platformdirs==4.4.0
pluggy==1.6.0
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.10.1
PyNaCl==1.6.0
pyOpenSSL==25.1.0
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.2
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
PyYAML==6.0.2
pyzmq==27.1.0
qrcode==8.2
realesrgan==0.3.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.9.0
simsimd==6.5.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
spake2==0.9
stack-data==0.6.3
starlette==0.48.0
stringzilla==4.2.1
supervisor==4.2.1
sympy==1.13.1
tb-nightly==2.21.0a20251014
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.1.0
Twisted==25.5.0
txaio==25.6.1
txtorcon==24.8.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
typing-inspection==0.4.2
uri-template==1.3.0
urllib3==2.5.0
uv==0.8.17
uvicorn==0.37.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.37.1
widgetsnbextension==4.0.14
yapf==0.43.0
zipp==1.0.0
zipstream-ng==1.9.0
zope.interface==7.2
